JAMES-2521 ReIndexer shouldhave its own tool module
