+++ mailbox/tools/pom.xml
@@ -32,6 +32,7 @@
     <name>Apache James :: Mailbox :: Tools</name>
 
     <modules>
+        <module>indexer</module>
         <module>jpa-migrator</module>
         <module>maildir-utils</module>
     </modules>
